Moved the class/function files to lib/utility + fixed issue with gh actions
